Small fix.
If you are interested then thanks to this fix on each execution we are gonna save 8 BYTES JUST ON ONE SINGLE VARIABLE! Isn't that great? Just think about it, 8 fucken bytes. For whatever we want it. It's ours. What are we gonna do with that living space? Future will tell.